work toward issue #124
changed the cast in the toLily method of text class to be a float rather than an int: this was causing some issues with testcase 31a because font sizes like 10.25 can't be cast by an int, but they're not text and yet the exception handler tells the system to output it as is rather than attaching "abs-fontsize". I think this has now fixed that particular testcase or whatever else I was working on has, but I'm waiting on the lilypond server to come back up so I can check. Also there's still some barcheck failures that I need to get rid of and a warning about a text spanner.

+++ implementation/primaries/Drawing/classes/Directions.py
@@ -30,7 +30,7 @@ class Text(BaseClass.Base):
         lilystring = ""
         if hasattr(self, "size"):
             try:
-                size = int(self.size)
+                size = float(self.size)
                 lilystring += "\\abs-fontsize #" + str(self.size) + " "
             except:
                 lilystring += "\\"+str(self.size)+" "
